+++ src/LandingPage/TeamLandingPage.tsx
@@ -2,20 +2,18 @@ import { useEffect, useState } from "react";
 import { useAuthStore } from "../context/AuthContext";
 import { useSimFBAStore } from "../context/SimFBAContext";
 import { useSimHCKStore } from "../context/SimHockeyContext";
-import { SectionCards } from "../_design/SectionCards";
-import { Text } from "../_design/Typography";
-import { getTextColorBasedOnBg } from "../_utility/getBorderClass";
-import { StandingsTable } from "../components/Common/Tables";
 import { Border } from "../_design/Borders";
 import { getLandingCFBData, getLandingNFLData } from "./TeamLandingPageHelper";
 import * as Titles from "./TeamLandingPageTitles";
 import { GetCurrentWeek } from "../_helper/teamHelper";
-import { Logo } from "../_design/Logo";
-import { Button } from "../_design/Buttons";
 import { LeagueType } from "./TeamLandingPageTitles";
-import { GamesBar,
-         TeamOverview
-       } from "./TeamLandingPageElements";
+import { GamesBar, 
+         TeamOverview,
+         TeamStandings, 
+         TeamMatchUp,
+         TeamMailbox, 
+         TeamStats } 
+         from "./TeamLandingPageComponents";
 
 interface TeamLandingPageProps {
   team: any;
@@ -27,7 +25,6 @@ export const TeamLandingPage = ({ team, league, ts }: TeamLandingPageProps) => {
   const { currentUser } = useAuthStore();
   const backgroundColor = team?.ColorOne || "#4B5563";
   const borderColor = team?.ColorTwo || "#4B5563";
-  const textColorClass = getTextColorBasedOnBg(backgroundColor);
   const { collegeNotifications, proNotifications, 
           allCFBStandings, allProStandings, 
           allCollegeGames, allProGames,
@@ -68,290 +65,127 @@ export const TeamLandingPage = ({ team, league, ts }: TeamLandingPageProps) => {
       break;
   }
 
-  console.log('roster data: ', rosterData)
-
   return (
     <>
       <div className="flex-col">
-        <div className="flex pb-1">
-          <div className="flex w-[72em] justify-start">
-            <GamesBar games={teamSchedule} league={league} 
-                      team={team} ts={ts} currentUser={currentUser} 
-                      backgroundColor={backgroundColor} 
-                      borderColor={borderColor} />
-          </div>
-        </div>
+        <GamesBar games={teamSchedule} 
+                  league={league} 
+                  team={team} 
+                  ts={ts} 
+                  currentUser={currentUser} 
+                  backgroundColor={backgroundColor} 
+                  borderColor={borderColor} 
+        />
         <div className="flex gap-4 items-start justify-center">
-        <Border
-            classes="border-4 py-0 px-0"
-            styles={{
-              backgroundColor: borderColor,
-              borderColor: backgroundColor,
-            }}
-        >
-          {currentUser && (
-            <SectionCards
-              team={team}
-              header={`${team.Conference} Standings`}
-              classes={`${textColorClass}, h-full`}
+          <Border
+              classes="border-4 py-0 px-0"
+              styles={{
+                backgroundColor: borderColor,
+                borderColor: backgroundColor,
+              }}
+          >
+            {currentUser && (
+              <TeamStandings standings={teamStandings}
+                             team={team}
+                             league={league}
+                             currentUser={currentUser}
+                             isLoadingTwo={isLoadingTwo}
+                             backgroundColor={backgroundColor}
+                             borderColor={borderColor}
+              />
+            )}
+          </Border>
+          <div className="flex flex-col items-center justify-center">
+            <Border
+              classes="border-4 py-[0px] px-[0px] w-[30em] h-[17em]"
+              styles={{
+                backgroundColor: borderColor,
+                borderColor: backgroundColor,
+              }}
             >
-              {isLoadingTwo ? (
-                <div className="flex justify-center items-center">
-                  <Text variant="small" 
-                        classes={`${textColorClass}`}>
-                    Loading...
-                  </Text>
-                </div>
-              ) : (
-                <StandingsTable standings={teamStandings} 
-                                league={league} 
-                                team={team} 
-                                currentUser={currentUser} />
-              )}
-            </SectionCards>
-          )}
-        </Border>
-        <div className="flex flex-col items-center justify-center">
-          <Border
-            classes="border-4 py-[0px] px-[0px] w-[30em] h-[17em]"
-            styles={{
-              backgroundColor: borderColor,
-              borderColor: backgroundColor,
-            }}
-          >
-            <SectionCards team={team} 
-                          header={`Week ${currentWeek} Match-Up`} 
-                          classes={`${textColorClass}`}>
-              {isLoadingTwo ? (
-                <div className="flex justify-center items-center">
-                  <Text variant="small" 
-                        classes={`${textColorClass}`}>
-                    Loading...
-                  </Text>
-                </div>
-              ) : teamMatchUp.length > 0 ? (
-                <>
-                  <div className="flex justify-center">
-                    <div className="flex-col">
-                      <Logo variant="normal" url={homeLogo}></Logo>
-                      <Text variant="small" 
-                            classes={`${textColorClass} 
-                                      font-semibold`} 
-                            className="pr-1">
-                        {homeLabel}
-                      </Text>
-                    </div>
-                    <Text variant="small" 
-                          classes={`${textColorClass} 
-                                    self-center 
-                                    font-semibold`}>
-                      vs
-                    </Text>
-                    <div className="flex-col">
-                      <Logo variant="normal" 
-                            url={awayLogo}>
-                      </Logo>
-                      <Text variant="small" 
-                            classes={`${textColorClass} font-semibold`} 
-                            className="pl-1">
-                        {awayLabel}
-                      </Text>
-                    </div>
-                  </div>
-                  <div className="flex-col items-center">
-                    <Text variant="small">
-                      {teamMatchUp[0].IsConference ? (teamMatchUp[0].IsDivisional ? 
-                            "Conference Divisional Game" : 
-                            "Conference Game") : 
-                            "Non-Conference Game"}
-                    </Text>
-                    <div className="flex justify-center gap-2 pt-1">
-                      <Button variant="primary" size="sm">Depth Chart</Button>
-                      <Button variant="primary" size="sm">Gameplan</Button>
-                    </div>
-                  </div>
-                </>
-              ) : (
-                <Text variant="small" classes={`${textColorClass} pt-2`}>
-                  <img src="./byeWeek.png" 
-                       alt="byeWeek" 
-                       className="max-h-[5em] justify-self-center"></img>
-                  < br></br>You don't have a game this week.< br></br>
-                  Enjoy the week off... go outside and touch some grass you degen.
-                </Text>
-              )}
-            </SectionCards>
-          </Border>
-          <Border
-            classes="border-4 py-[0px] px-[0px] w-[30em] h-[22em]"
-            styles={{
-              backgroundColor: borderColor,
-              borderColor: backgroundColor,
-            }}
-          >
-            <SectionCards
-              team={team}
-              header="Team Overview"
-              classes={`${textColorClass}`}
+              <TeamMatchUp team={team}
+                           week={currentWeek}
+                           matchUp={teamMatchUp}
+                           homeLogo={homeLogo}
+                           awayLogo={awayLogo}
+                           homeLabel={homeLabel}
+                           awayLabel={awayLabel}
+                           backgroundColor={backgroundColor}
+                           isLoadingTwo={isLoadingTwo}
+              />
+            </Border>
+            <Border
+              classes="border-4 py-[0px] px-[0px] w-[30em] h-[22em]"
+              styles={{
+                backgroundColor: borderColor,
+                borderColor: backgroundColor,
+              }}
             >
-              {isLoadingTwo ? (
-                <div className="flex justify-center items-center">
-                  <Text variant="small" classes={`${textColorClass}`}>
-                    Loading...
-                  </Text>
-                </div>
-              ) : (
-                <div className="p-1">
-                  <TeamOverview
-                    team={team}
-                    league={league}
-                    rosterData={rosterData}
-                    ts={ts}
-                    currentUser={currentUser}
-                    backgroundColor={backgroundColor}
-                    borderColor={borderColor}
-                  />
-                </div>
-              )}
-            </SectionCards> 
-          </Border>
-          <Border
-            classes="border-4 py-[0px] px-[0px] w-[30em] h-[12em]"
-            styles={{
-              backgroundColor: borderColor,
-              borderColor: backgroundColor,
-            }}
-          >
-            <SectionCards team={team} header="Inbox" 
-                          classes={`${textColorClass}`}>
-              {isLoadingTwo ? (
-                <div className="flex justify-center items-center">
-                  <Text variant="small" 
-                        classes={`${textColorClass}`}>
-                    Loading...
-                  </Text>
-                </div>
-              ) : teamNotifications.length > 0 ? (
-                teamNotifications.map((notification, index) => (
-                  <div key={index} className="mb-2">
-                    <Text variant="small" classes={`${textColorClass}`}>
-                      {notification.Subject}
-                    </Text>
-                    <Text variant="small" classes={`${textColorClass}`}>
-                      {notification.Message}
-                    </Text>
-                  </div>
-                ))
-              ) : (
-                <Text variant="small" classes={`${textColorClass}`}>
-                  Your Inbox is Empty
-                </Text>
-              )}
-            </SectionCards>
-          </Border>
-        </div>
-        <div className="flex flex-col items-center justify-center gap-1">
-        <Border
-            classes="border-4 py-[0px] px-[0px] w-[25em] h-[15em]"
-            styles={{
-              backgroundColor: borderColor,
-              borderColor: backgroundColor,
-            }}
-          >
-            <SectionCards team={team} header={headers[0]} 
-                          classes={`${textColorClass}`}>
-              {isLoadingTwo ? (
-                <div className="flex justify-center items-center">
-                  <Text variant="small" classes={`${textColorClass}`}>
-                    Loading...
-                  </Text>
-                </div>
-              ) : teamNotifications.length > 0 ? (
-                teamNotifications.map((notification, index) => (
-                  <div key={index} className="mb-2">
-                    <Text variant="small" 
-                          classes={`${textColorClass}`}>
-                      {notification.Subject}
-                    </Text>
-                    <Text variant="small" 
-                          classes={`${textColorClass}`}>
-                      {notification.Message}
-                    </Text>
-                  </div>
-                ))
-              ) : (
-                <Text variant="small" classes={`${textColorClass}`}>
-                  No Stats to Show
-                </Text>
-              )}
-            </SectionCards>
-          </Border>
-          <Border
-            classes="border-4 py-[0px] px-[0px] w-[25em] h-[15em]"
-            styles={{
-              backgroundColor: borderColor,
-              borderColor: backgroundColor,
-            }}
-          >
-            <SectionCards team={team} header={headers[1]} 
-                          classes={`${textColorClass}`}>
-              {isLoadingTwo ? (
-                <div className="flex justify-center items-center">
-                  <Text variant="small" classes={`${textColorClass}`}>
-                    Loading...
-                  </Text>
-                </div>
-              ) : teamNotifications.length > 0 ? (
-                teamNotifications.map((notification, index) => (
-                  <div key={index} className="mb-2">
-                    <Text variant="small" classes={`${textColorClass}`}>
-                      {notification.Subject}
-                    </Text>
-                    <Text variant="small" classes={`${textColorClass}`}>
-                      {notification.Message}
-                    </Text>
-                  </div>
-                ))
-              ) : (
-                <Text variant="small" classes={`${textColorClass}`}>
-                  No Stats to Show
-                </Text>
-              )}
-            </SectionCards>
-          </Border>
-          <Border
-            classes="border-4 py-[0px] px-[0px] w-[25em] h-[15em]"
-            styles={{
-              backgroundColor: borderColor,
-              borderColor: backgroundColor,
-            }}
-          >
-            <SectionCards team={team} header={headers[2]} 
-                          classes={`${textColorClass}`}>
-              {isLoadingTwo ? (
-                <div className="flex justify-center items-center">
-                  <Text variant="small" classes={`${textColorClass}`}>
-                    Loading...
-                  </Text>
-                </div>
-              ) : teamNotifications.length > 0 ? (
-                teamNotifications.map((notification, index) => (
-                  <div key={index} className="mb-2">
-                    <Text variant="small" classes={`${textColorClass}`}>
-                      {notification.Subject}
-                    </Text>
-                    <Text variant="small" classes={`${textColorClass}`}>
-                      {notification.Message}
-                    </Text>
-                  </div>
-                ))
-              ) : (
-                <Text variant="small" classes={`${textColorClass}`}>
-                  No Stats to Show
-                </Text>
-              )}
-            </SectionCards>
-          </Border>
+                <TeamOverview
+                  team={team}
+                  league={league}
+                  rosterData={rosterData}
+                  ts={ts}
+                  currentUser={currentUser}
+                  backgroundColor={backgroundColor}
+                  borderColor={borderColor}
+                  isLoadingTwo={isLoadingTwo}
+                />
+            </Border>
+            <Border
+              classes="border-4 py-[0px] px-[0px] w-[30em] h-[12em]"
+              styles={{
+                backgroundColor: borderColor,
+                borderColor: backgroundColor,
+              }}
+            >
+              <TeamMailbox team={team}
+                           notifications={teamNotifications}
+                           backgroundColor={backgroundColor}
+                           isLoadingTwo={isLoadingTwo}
+              />
+            </Border>
+          </div>
+          <div className="flex flex-col items-center justify-center gap-1">
+            <Border
+                classes="border-4 py-[0px] px-[0px] w-[25em] h-[15em]"
+                styles={{
+                  backgroundColor: borderColor,
+                  borderColor: backgroundColor,
+                }}
+              >
+                <TeamStats team={team}
+                          header={headers[0]}
+                          backgroundColor={backgroundColor}
+                          isLoadingTwo={isLoadingTwo}
+                />
+              </Border>
+              <Border
+                classes="border-4 py-[0px] px-[0px] w-[25em] h-[15em]"
+                styles={{
+                  backgroundColor: borderColor,
+                  borderColor: backgroundColor,
+                }}
+              >
+                <TeamStats team={team}
+                          header={headers[1]}
+                          backgroundColor={backgroundColor}
+                          isLoadingTwo={isLoadingTwo}
+                />
+              </Border>
+              <Border
+                classes="border-4 py-[0px] px-[0px] w-[25em] h-[15em]"
+                styles={{
+                  backgroundColor: borderColor,
+                  borderColor: backgroundColor,
+                }}
+              >
+                <TeamStats team={team}
+                          header={headers[2]}
+                          backgroundColor={backgroundColor}
+                          isLoadingTwo={isLoadingTwo}
+                />
+              </Border>
           </div>
         </div>
       </div>
